% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{$.MVL}
\alias{$.MVL}
\title{MVL handle subscription operator}
\usage{
\method{$}{MVL}(MVLHANDLE, name)
}
\arguments{
\item{MVLHANDLE}{- handle to opened MVL file as generated by \code{mvl_open}}

\item{name}{- name of object to retrieve}
}
\value{
Stored object
}
\description{
Retrieve objects stored in the library. Unlike for R lists the match on name is always exact.
}
