% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{dim.MVL_OBJECT}
\alias{dim.MVL_OBJECT}
\title{Obtain dimensions of MVL object}
\usage{
\method{dim}{MVL_OBJECT}(x)
}
\arguments{
\item{x}{MVL_OBJECT as retrieved by subscription operators}
}
\value{
object dimensions, or NULL if not present
}
\description{
Obtain dimensions of MVL object
}
