% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{length.MVL_OBJECT}
\alias{length.MVL_OBJECT}
\title{Obtain length of MVL object}
\usage{
\method{length}{MVL_OBJECT}(x)
}
\arguments{
\item{x}{MVL_OBJECT as retrieved by subscription operators}
}
\value{
object length as stored in MVL file. This is the total length of object for arrays, and number of columns for data frames.
}
\description{
Obtain length of MVL object
}
