% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl2R}
\alias{mvl2R}
\title{Make sure the object is fully converted to its R representation}
\usage{
mvl2R(obj, raw = FALSE)
}
\arguments{
\item{obj}{- MVL object retrieved by subscription of MVL library or other objects}

\item{raw}{- request to return data in raw format when it does not map exactly to R data types.}
}
\value{
Stored object
}
\description{
If the object is stored in MVL file, we return its pure R representation. 
Otherwise, we return the object itself.
}
