% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_class}
\alias{mvl_class}
\title{Return underlying R class of object}
\usage{
mvl_class(x)
}
\arguments{
\item{x}{any object}
}
\value{
character string giving object class
}
\description{
This function returns the class of underlying MVL_OBJECT, i.e. the class it would have if converted into regular R object.
For non-MVL_OBJECTs it simply calls usual R class()
}
