% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_get_neighbors}
\alias{mvl_get_neighbors}
\title{Retrieve indices of nearby rows.}
\usage{
mvl_get_neighbors(spatial_index, data_list)
}
\arguments{
\item{spatial_index}{MVL_OBJECT computed by \code{mvl_write_spatial_groups}}

\item{data_list}{a list of vectors of equal length. They can be MVL_OBJECTs or R vectors.}
}
\value{
a list of vectors of indices
}
\description{
This function is passed the index computed by \code{mvl_write_spatial_groups} and a list of vectors, which rows are interpreted as points.
For each row, the function returns a list of indices describing rows that are close to it.
}
\seealso{
\code{\link{mvl_write_spatial_groups}}
}
