% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_group_lapply}
\alias{mvl_group_lapply}
\title{Apply function to index stretches}
\usage{
mvl_group_lapply(G, fn)
}
\arguments{
\item{G}{a list of groups and group stretches produced by \code{mvl_group}}

\item{fn}{a function of one argument - list of indices}
}
\value{
a list of results of function \code{fn}
}
\description{
Iteratively call function \code{fn} over index stretches previously computed with \code{mvl_group}
}
\seealso{
\code{\link{mvl_group}}
}
