% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_merge}
\alias{mvl_merge}
\title{Merge two MVL data frames and write the result}
\usage{
mvl_merge(
  MVLHANDLE,
  df1,
  df2,
  name = NULL,
  by = NULL,
  by.x = by,
  by.y = by,
  suffixes = c(".x", ".y"),
  only.columns.x = NULL,
  only.columns.y = NULL
)
}
\arguments{
\item{MVLHANDLE}{a handle to MVL file produced by \code{mvl_open()}}

\item{df1}{a data.frame stored in MVL file}

\item{df2}{a data.frame stored in MVL file}

\item{name}{if specified add a named entry to MVL file directory}

\item{by}{list of columns to use as key}

\item{by.x}{list of columns to use as key for \code{df1}}

\item{by.y}{list of columns to use as key for \code{df1}}

\item{suffixes}{rename columns with identical names using these suffixes}

\item{only.columns.x}{only copy these columns from df1}

\item{only.columns.y}{only copy these columns from df2}
}
\description{
Merge two MVL data frames and write the result
}
