% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_verify}
\alias{mvl_verify}
\title{Verify object integrity.}
\usage{
mvl_verify(x)
}
\arguments{
\item{x}{MVL handle or MVL object}
}
\description{
To verify entire MVL file pass an open handle. Otherwise pass MVL_OBJECT. 
Verification will fail with an error if the checksums do not match or are not available
}
