% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_xlength}
\alias{mvl_xlength}
\title{Return length of R vector in numeric format}
\usage{
mvl_xlength(x)
}
\arguments{
\item{x}{any R object}
}
\value{
length of object in numeric format
}
\description{
Internally this calls R function xlength() rather than length(). This allows to obtain length of larger vectors
}
