% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{names.MVL}
\alias{names.MVL}
\title{Print MVL directory}
\usage{
\method{names}{MVL}(x)
}
\arguments{
\item{x}{handle to MVL file as created by \code{mvl_open}}
}
\value{
character vector of names present in the directory
}
\description{
Print MVL directory
}
