% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{names.MVL_OBJECT}
\alias{names.MVL_OBJECT}
\title{Retrieve MVL object names}
\usage{
\method{names}{MVL_OBJECT}(x)
}
\arguments{
\item{x}{MVL_OBJECT as retrieved by subscription operators}
}
\value{
character vector of names
}
\description{
Retrieve MVL object names
}
