% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{print.MVL}
\alias{print.MVL}
\title{Print MVL}
\usage{
\method{print}{MVL}(x, ...)
}
\arguments{
\item{x}{handle to MVL file as created by \code{mvl_open}}

\item{\ldots}{not used.}
}
\value{
invisible(MVLHANDLE)
}
\description{
Print MVL
}
