% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{print.MVL_INDEX}
\alias{print.MVL_INDEX}
\title{Print summary information of MVL_INDEX}
\usage{
\method{print}{MVL_INDEX}(x, ...)
}
\arguments{
\item{x}{MVL_INDEX object}

\item{\ldots}{not used.}
}
\value{
invisible(obj)
}
\description{
Print summary information of MVL_INDEX
}
