% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{print.MVL_OBJECT}
\alias{print.MVL_OBJECT}
\title{Print MVL object
This is a convenience function for displaying MVL_OBJECTs.}
\usage{
\method{print}{MVL_OBJECT}(x, ..., small_length = 10)
}
\arguments{
\item{x}{MVL_OBJECT as retrieved by subscription operators}

\item{small_length}{do not list more than this number of columns in data frames}

\item{\ldots}{not used.}
}
\value{
invisible(obj)
}
\description{
Print MVL object
This is a convenience function for displaying MVL_OBJECTs.
}
