% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{[.MVL_OBJECT}
\alias{[.MVL_OBJECT}
\title{MVL object subscription operator}
\usage{
\method{[}{MVL_OBJECT}(obj, i, ..., drop = TRUE, raw = FALSE, recurse = FALSE, ref = FALSE)
}
\arguments{
\item{obj}{- MVL object retrieved by subscription of MVL library or other objects}

\item{i}{- optional index.}

\item{drop}{- whether to drop dimensionality, such as done with R array or data frames}

\item{raw}{- request to return data in raw format when it does not map exactly to R data types.}

\item{recurse}{- force recursive conversion to pure R objects.}

\item{ref}{- always return an MVL_OBJECT}

\item{\ldots}{optional additional indices for multidimensional arrays and data frames}
}
\value{
Stored object
}
\description{
Retrieve objects stored in mappable vector library. Large nested objects are returned as instances of MVL_OBJECT to delay access until needed.
See \code{mvl_open} for example.
}
