% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_class}
\alias{mvl_class}
\title{Return underlying R class of object}
\usage{
mvl_class(x)
}
\arguments{
\item{x}{any object}
}
\value{
character string giving object class
}
\description{
This function returns the equivalent R class of underlying MVL object, i.e. the class it would have if converted into a regular R object.
For non-MVL objects the function simply calls the usual R class(), so it can be used instead of class() for code that operates on both usual R objects and MVL objects.
}
