% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_get_groups}
\alias{mvl_get_groups}
\title{Retrieve indices belong to one or more groups}
\usage{
mvl_get_groups(prev, first_indices)
}
\arguments{
\item{prev}{MVL_OBJECT \code{prev} computed by \code{mvl_write_groups}}

\item{first_indices}{indices from \code{first} vector computed by \code{mvl_write_groups}}
}
\value{
a vector of indices
}
\description{
This function is passed the \code{prev} vector computed by \code{mvl_write_groups} and one or more indices from the \code{first} vector.
}
\seealso{
\code{\link{mvl_group}}
}
