% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMVL.R
\name{mvl_indexed_copy}
\alias{mvl_indexed_copy}
\title{Index copy vector}
\usage{
mvl_indexed_copy(MVLHANDLE, x, indices, name = NULL, only.columns = NULL)
}
\arguments{
\item{MVLHANDLE}{a handle to MVL file produced by mvl_open()}

\item{x}{a suitable R object (vector, array, list, data.frame) or a vector-like MVL_OBJECT}

\item{indices}{list of indices into x}

\item{name}{if specified add a named entry to MVL file directory}

\item{only.columns}{if x is MVL_OBJECT with class data.frame copy only columns specified in this character or integer vector}
}
\description{
Index copy vector
}
